Restructure folders and files in pages
