+++ src/pages/homepage/cards/CardInfo.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
-import eggshell from "../../assets/eggshell.png";
+import eggshell from "../../../assets/eggshell.png";
 import { Link } from "react-router-dom";
-import { members } from "../../data/members";
+import { members } from "../../../data/members";
 
 const Container = styled.div`
   background-image: url(${eggshell});
